refactor(jest-preset): extract importComponent plugin into file
